Export all tracks

+++ qcard-tools/MidiQcard/QcardToMidi/Program.cs
@@ -1,31 +1,57 @@
 using System.Diagnostics;
 using QcardToMidi;
 
-Debug.Assert(args.Length >= 3, "<path to qcard> <song number> <output file>");
-string inputPath = args[0];
-int trackNum = int.Parse(args[1]) - 1;
-Debug.Assert(trackNum >= 0);
-string outputPath = args[2];
+if (args.Length < 1)
+{
+    Console.Out.WriteLine("<path to qcard> [song number] [output file]\n" +
+                          " Output file can be .bin for raw midi events, .mid/.midi for a midi file\n" +
+                          " Omit output file and all tracks will be generated.");
+    return;
+}
 
+string inputPath = args[0];
 QCard qCard = new(File.ReadAllBytes(inputPath));
 Console.WriteLine(qCard);
 
-if (File.Exists(outputPath)) File.Delete(outputPath);
-using FileStream fileStream = File.Create(outputPath);
-using BinaryWriter writer = new(fileStream);
-
-switch (Path.GetExtension(outputPath).ToLower().TrimStart('.'))
+if (args.Length >= 3)
 {
-    case "mid":
-    case "midi":
-        qCard.ConvertToMidiFile(writer, trackNum);
-        Console.WriteLine($"Exported MIDI to {outputPath}");
-        break;
-    case "bin":
-        qCard.ConvertToMidiStream(writer, trackNum, writeTimes: false, muteSpecials: true);
-        Console.WriteLine($"Exported raw MIDI event data to {outputPath}");
-        break;
-    default:
-        throw new ArgumentException("Unsupported output file type, supports .mid/.midi or .bin");
+    int trackNum = int.Parse(args[1]) - 1;
+    Debug.Assert(trackNum >= 0);
+    string outputPath = args[2];
+
+
+    if (File.Exists(outputPath)) File.Delete(outputPath);
+    using FileStream fileStream = File.Create(outputPath);
+    using BinaryWriter writer = new(fileStream);
+
+    switch (Path.GetExtension(outputPath).ToLower().TrimStart('.'))
+    {
+        case "mid":
+        case "midi":
+            qCard.ConvertToMidiFile(writer, trackNum);
+            Console.WriteLine($"Exported MIDI to {outputPath}");
+            break;
+        case "bin":
+            qCard.ConvertToMidiStream(writer, trackNum, writeTimes: false, muteSpecials: true);
+            Console.WriteLine($"Exported raw MIDI event data to {outputPath}");
+            break;
+        default:
+            throw new ArgumentException("Unsupported output file type, supports .mid/.midi or .bin");
+    }
+
+    return;
 }
 
+if (args.Length >= 1)
+{
+    string dir = Path.GetDirectoryName(inputPath);
+    string basename = Path.GetFileNameWithoutExtension(inputPath);
+    for (int i = 0; i < qCard.TrackCount; i++)
+    {
+        string outputPath = Path.Combine(dir, $"{basename}_track{i+1:d2}.mid");
+        using FileStream fileStream = File.Create(outputPath);
+        using BinaryWriter writer = new(fileStream);
+        qCard.ConvertToMidiFile(writer, i);
+        Console.WriteLine($"Exported {Path.GetFileName(outputPath)}");
+    }
+}
